% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_g2covs_all.R
\name{g2c_download_all_covs}
\alias{g2c_download_all_covs}
\title{Function to download all geospatial covariates for a year}
\usage{
g2c_download_all_covs(
  ISO,
  year,
  output_dir,
  rst_mask = NULL,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{ISO}{A character, ISO code(s) of a specified country}

\item{year}{Numerical, the year of the covariate}

\item{output_dir}{Character vector containing the path to the directory for
writing output files. Default is the temp directory.}

\item{rst_mask}{path to mask file (Default NULL)}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar.}

\item{...}{Additional arguments:\cr
\code{ftp_srv}: Boolean, If TRUE then WorldPop ftp server will
be used otherwise http.\cr
\code{method}: Method to be used for downloading files. Current download
methods are "internal", "wininet" (Windows only) "libcurl",  "wget" and "curl",
and there is a value "auto"
\code{cells_touches}:     logical. If TRUE, all cells touched by lines or ,
polygons are affected, not just those on the line render path
or whose center point is within the polygon.\cr
\code{keep_tmp}:     logical. If TRUE, do not remove ,
any temporal files downloaded.\cr}
}
\value{
A character, path to downloaded file
}
\description{
Function to download all geospatial covariates for a year
}
\examples{
\dontrun{
g2c_download_all_covs(ISO=NULL,
                      year=2021,
                      rst_mask=NULL,
                      output_dir=NULL,
                      quiet=TRUE)
}
}
