% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_covariates.R
\name{get_names_covariates}
\alias{get_names_covariates}
\title{Functions to get all names of the available covaraites}
\usage{
get_names_covariates()
}
\value{
List, List of the names of the available covaraites
}
\description{
Functions to get all names of the available covaraites
}
\examples{
\dontrun{
get_names_covariates()
}
}
